pyproject.toml
[project]
name = "agents-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "openai-agents>=0.0.6",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]
